requirements.txt
pyyaml
requests
click
deepdiff
pylint
